hotfix - reestruturando o cardpokemon
