+++ package.json
@@ -21,6 +21,7 @@
     "react-scripts": "5.0.0",
     "react-spinners": "^0.11.0",
     "redux": "^4.1.2",
+    "redux-persist": "^6.0.0",
     "redux-saga": "^1.1.3",
     "styled-components": "^5.3.3",
     "web-vitals": "^2.1.4"
