+++ src/index.js
@@ -1,14 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { PersistGate } from 'redux-persist/integration/react';
 import { Provider } from 'react-redux';
 import App from './App';
 
-import store from './redux/store';
+import { store, persistor } from './redux/store';
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store} >
-      <App />
+      <PersistGate loading={null} persistor={persistor} >
+        <App />
+      </PersistGate>
     </Provider>
   </React.StrictMode>,
   document.getElementById('root')
