+++ src/pages/Home/styles.js
@@ -5,3 +5,17 @@ export const Container = styled.div`
   width: 100vw;
   background-color: #E5E5E5;
 `;
+
+export const Button = styled.button`
+  background-color: #F2243A;
+  color: #FFF;
+
+  width: 200px;
+  height: 50px;
+  margin-bottom: 20px;
+  font-size: 16px;
+  font-weight: bold;
+
+  border: none;
+  border-radius: 20px;
+`;
