+++ src/redux/store/index.js
@@ -1,13 +1,23 @@
 import { createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
+import { persistStore, persistReducer } from 'redux-persist';
+import storage from 'redux-persist/lib/storage';
 
 import rootReducers from '../reducers/index';
 import rootSaga from '../sagas/pokemon.saga';
 
+const persistConfig = {
+  key: 'root',
+  storage,
+};
+
+const persistedReducer = persistReducer(persistConfig, rootReducers);
+
 const sagaMiddleware = createSagaMiddleware();
 
-const store = createStore(rootReducers, applyMiddleware(sagaMiddleware));
+const store = createStore(persistedReducer, applyMiddleware(sagaMiddleware));
+const persistor = persistStore(store);
 
-export default store;
+export { store, persistor };
 
 sagaMiddleware.run(rootSaga);
